--- Software/misc/sine_wave.csv
@@ -1,21 +1,0 @@
-Time,Signal
-0.0,0.0
-0.05,2.939
-0.1,4.755
-0.15,4.755
-0.2,2.939
-0.25,0.0
-0.3,-2.939
-0.35,-4.755
-0.4,-4.755
-0.45,-2.939
-0.5,-0.0
-0.55,2.939
-0.6,4.755
-0.65,4.755
-0.7,2.939
-0.75,0.0
-0.8,-2.939
-0.85,-4.755
-0.9,-4.755
-0.95,-2.939
