--- Software/misc/telemetry.csv
@@ -1,2 +1,0 @@
-Timestamp: 2024-12-06 22:39:52
-Timestamp_HMS,PubName/format,STATE/PITCH,STATE/ROLL,STATE/LW,STATE/RW,STATE/LINVEL,STATE/ANGVEL,STATE/XPOS,STATE/YPOS,STATE/THETA,STATE/TL,STATE/TR,STATE/UL,STATE/UR,STATE/UB,STATE/UA,STATE/UV,IMURAW/AX,IMURAW/AY,IMURAW/AZ,IMURAW/GX,IMURAW/GY,IMURAW/GZ
