+++ Software/scripts/deployServices.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
+# ----------------------------------
+#  @file    deployServices.sh
+#  @brief   Updates and Deploys the services to the system.
 
 echo "[DEPLOY_SERVICES] Available:"
 
@@ -15,16 +18,11 @@ sudo cp /home/pi/prog/software/services/twsb.service /etc/systemd/system/
 sudo cp /home/pi/prog/software/services/joystick.service /etc/systemd/system/
 sudo cp /home/pi/prog/software/services/vision.service /etc/systemd/system/
 
-
 # Reload systemd to pick up new/updated units
 sudo systemctl daemon-reload
 
-# Enable & start twsb.service and joystick.service
 sudo systemctl enable --now twsb.service
 sudo systemctl enable --now joystick.service
 sudo systemctl enable --now vision.service
-
-
-# Conditionally enable/start cam or vision
 
 echo "[DEPLOY_SERVICE] complete."
