--- Software/src/main.cpp
@@ -1,9 +1,0 @@
-#include "iostream"
-
-int main(int argc, char *argv[]) {
-
-
-
-    
-    return 0;
-}
